requirements.txt
azure-identity
openai
python-dotenv
fastapi
uvicorn[standard]
